fix: loading assets without having to load entire asset bundles
This allows us to dynamically load assets from an asset bundle without having to load the entire contents of the asset bundle into memory first. This is huge in terms of time saving when an application opens initially as it doesn't have to fetch and parse entire bundles which can take several seconds depending on the bundle size, and is much more performant and versatile. Assets fetched this way are added to the bundle's run-time cache and are fetched much more quickly on subsequent fetches too.