feat/fix: abstract away main application into an interface to prevent circular dependencies between dlls
This is actually pretty huge, as I had my overall project architecture somewhat like a ball of yarn at times; lib-platform would need access to lib-core, but lib-core needed specific access to one part of lib-platform, but that can't be done or you fall into the rabbit hole of circular dependencies.

This entire commit fixes that. We abstract away the `Application` class into an interface and we expose `IApplication` to any DLL that needs it. Doing it this way allows ALL of our DLLs to access something crucial: the `App` pointer, which holds a ton of useful things, like the bundle manager and audio manager.

This commit also introduces some Lib-Renderer specific code; primarily some shaders that the map uses, support for a camera (included frustum culling!) and dedicated classes for handling shaders, parsing map files, and rendering map files.