fix: using assimp in App-* executables; also cleans up the premake configs for all App-* projects
We were dynamically linking against Assimp which requires the assimp-* dll in the exe directory. I didn't want that and so I fixed the configurations to instead statically link against it.

This commit also cleans up and further organizes the premake configurations for every App-* and a few Lib-*s. We now expect each project to handle copying directories and files separately instead of doing so in the main Build.lua config file. Keeps things cleaner and more concise than it used to be.