feat: ship compressed dependency libs and uncompress during the build pipeline
This was primarily due to the large file size of Assimp's static libs. They range from 100MB for dist builds to 300+MB for debug builds.. pretty huge.
I could have opted for other methods such as using Git LFS or hosting these on a remote server and fetching them during the build pipeline, but I chose not to. Requiring users to install and use Git LFS is not what I wanted, and hosting the files somewhere else could arguably cause trust issues which no sane dev would be alright with.

My solution is to compress each dependency's lib directory and its contents, ship those to the repo, and uncompress each of them when the build pipeline fires for the first time. This is better (in my opinion) as it keeps everything here in a single repository which avoids both of the things that I was against doing.