fix/feat: get rendering working and fix a bunch of invalid access errors when using GL functions across multiple DLLs
Tons of changes in this commit. We add a GL helper class to expose GL functions to all DLLs which prevents invalid access errors across the project.

A ton of rendering class code in Lib-Renderer has been modified so that we only ever pull asset data from the asset bundle instead of requiring individual asset access from separate files - as it should have been done in the first place.

We also ensure map caching via the MapManager class so we don't have to re-load maps we've already seen before, which is great for ensuring that we re-use map data instead of having to parse maps multiple times.